+++ packages/project-builder-web/src/utils/time-ago.ts
@@ -1,0 +1,12 @@
+import TimeAgo from 'javascript-time-ago';
+import en from 'javascript-time-ago/locale/en';
+
+let timeAgoInitialized = false;
+
+export function initializeTimeAgo(): void {
+  if (timeAgoInitialized) {
+    return;
+  }
+  TimeAgo.addDefaultLocale(en);
+  timeAgoInitialized = true;
+}